frozen_requirements.txt
antlr4-python3-runtime==4.7.2
appdirs==1.4.3
asteval==0.9.14
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cvxpy==1.0.23
cycler==0.10.0
dataclasses==0.6
decorator==4.4.0
dill==0.2.9
ecos==2.0.7.post1
forest-benchmarking==0.4.0
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
immutables==0.6
kiwisolver==1.1.0
lmfit==0.9.13
matplotlib==3.1.0
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multiprocess==0.70.7
networkx==2.3
numpy==1.16.3
osqp==0.5.0
pandas==0.24.2
pluggy==0.11.0
py==1.8.0
pyparsing==2.4.0
pyquil==2.8.0
pytest==4.5.0
python-dateutil==2.8.0
python-rapidjson==0.7.1
pytz==2019.1
pyzmq==18.0.1
requests==2.22.0
rpcq==2.6.1
ruamel.yaml==0.15.96
scipy==1.3.0
scs==2.1.0
six==1.12.0
smmap2==2.0.5
sympy==1.4
toml==0.10.0
tqdm==4.32.1
typing==3.6.6
uncertainties==3.1
urllib3==1.25.3
wcwidth==0.1.7
